Move enumeratee functions to contrib.scalaz
