+++ foundation/src/main/scala/quasar/contrib/scalaz/enumeratee.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package quasar.fp
+package quasar.contrib.scalaz
 
 import quasar.Predef._
